Make FACTx an int, more efficient than float.

+++ SimpleGifAnimViewer/SimpleGifAnimViewer.ino
@@ -6,8 +6,8 @@
 // while displaying it, or bouncing it around if it's smaller than the display
 int OFFSETX = 0;
 int OFFSETY = 0;
-float FACTX = 0;
-float FACTY = 0;
+int FACTX = 0;
+int FACTY = 0;
 
 #ifdef ESP8266
 #include <FS.h>
