+++ SimpleGifAnimViewer2/SimpleGifAnimViewer2.ino
@@ -8,8 +8,8 @@
 // while displaying it, or bouncing it around if it's smaller than the display
 int OFFSETX = 0;
 int OFFSETY = 0;
-float FACTX = 0;
-float FACTY = 0;
+int FACTX = 0;
+int FACTY = 0;
 
 
 const char *pathname = "/gifs64/200_circlesmoke.gif";
